+++ demo/src/app/components/+datepicker/disabled/datepicker-disabled.component.ts
@@ -1,0 +1,7 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-demo-datepicker-disabled',
+    templateUrl: './datepicker-disabled.component.html'
+})
+export class DemoDatePickerDisabledComponent {}