+++ src/mention/suggestions/suggestions.component.ts
@@ -1,9 +1,10 @@
-import { Component, OnInit, OnDestroy, HostBinding, ElementRef } from '@angular/core';
+import { Component, OnInit, OnDestroy, HostBinding, ElementRef, NgZone } from '@angular/core';
 import { Subject, Observable, of } from 'rxjs';
 import { ThySelectionListChange } from '../../list';
 import { MentionDefaultDataItem, Mention, MentionSuggestionSelectEvent } from '../interfaces';
-import { debounceTime, switchMap, catchError } from 'rxjs/operators';
+import { debounceTime, switchMap, catchError, take } from 'rxjs/operators';
 import { SeekQueryResult } from '../adapter/adapter';
+import { ThyPopoverRef } from '../../popover';
 
 @Component({
     selector: 'thy-mention-suggestions',
@@ -24,11 +25,15 @@ export class ThyMentionSuggestionsComponent<TItem = MentionDefaultDataItem> impl
 
     @HostBinding('class.thy-mention-suggestions') suggestionsClass = true;
 
-    constructor(public elementRef: ElementRef<HTMLElement>) {
+    constructor(
+        public elementRef: ElementRef<HTMLElement>,
+        private ngZone: NgZone,
+        private popoverRef: ThyPopoverRef<any>
+    ) {
         this.search$
             .pipe(
                 switchMap(query => {
-                    const data = this.filterData(query.term, this.mention);
+                    const data = this.mention.search(query.term, this.mention.data);
                     if (data instanceof Observable) {
                         this.loadingDone = false;
                         return (data as Observable<TItem[]>).pipe(debounceTime(this.debounce));
@@ -43,19 +48,17 @@ export class ThyMentionSuggestionsComponent<TItem = MentionDefaultDataItem> impl
             )
             .subscribe(data => {
                 this.loadingDone = true;
-                this.data = data;
-            });
-    }
+                this.data = data || [];
 
-    private filterData(term: string, mention: Mention<MentionDefaultDataItem>) {
-        const data = mention.data;
-        if (mention.search) {
-            return mention.search(term, data);
-        } else {
-            return data.filter(item => {
-                return !item.name || item.name.toLowerCase().includes(term.toLowerCase());
+                if (this.popoverRef) {
+                    if (this.mention.autoClose && this.data.length === 0) {
+                        this.popoverRef.close();
+                    }
+                    this.ngZone.onStable.pipe(take(1)).subscribe(() => {
+                        this.popoverRef.updatePosition();
+                    });
+                }
             });
-        }
     }
 
     ngOnInit(): void {